% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\docType{data}
\name{all_deaths}
\alias{all_deaths}
\title{all_deaths}
\format{A data frame}
\usage{
all_deaths
}
\description{
Game of Thrones Deaths
https://www.kaggle.com/mylesoneill/game-of-thrones#character-deaths.csv
}
\keyword{datasets}
