% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\docType{data}
\name{backpack}
\alias{backpack}
\title{Backpack Weight and Back Problems Data}
\format{
A data frame
}
\usage{
backpack
}
\description{
A dataset looking at the weight of backpack carried by higher education students and back problems.
}
\keyword{datasets}
