% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\docType{data}
\name{bank_expense}
\alias{bank_expense}
\title{bank_expense}
\format{
A data frame
}
\usage{
bank_expense
}
\description{
Selected items from large bank's profit/loss statement
}
\keyword{datasets}
