% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\docType{data}
\name{beer}
\alias{beer}
\title{Beer}
\format{A data frame}
\usage{
beer
}
\description{
A dataset with the brand, brewer, alcohol percentage,
 calories and grams of carhohydrates for 172 beers.
}
\details{
Some data cleaning required for this data.
}
\keyword{datasets}
