% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\docType{data}
\name{benefits}
\alias{benefits}
\title{Unemployement of Blue Collar Workers}
\format{A time serie containing :

stateur
state unemployment rate (in %)

statemb
state maximum benefit level

state
state of residence code

age
age in years

tenure
years of tenure in job lost

joblost
a factor with levels (slack_work,position_abolished,seasonal_job_ended,other)

nwhite
non-white ?

school12
more than 12 years of school ?

sex
a factor with levels (male,female)

bluecol
blue collar worker ?

smsa
lives is smsa ?

married
married ?

dkids
has kids ?

dykids
has young kids (0-5 yrs) ?

yrdispl
year of job displacement (1982=1,..., 1991=10)

rr
replacement rate

head
is head of household ?

ui
applied for (and received) UI benefits ?}
\source{
McCall, B.P. (1995) “The impact of unemployment insurance benefit levels on recipiency”, Journal of Business and Economic Statistics, 13, 189–198.
}
\usage{
benefits
}
\description{
a cross-section from 1972

number of observations : 4877

observation : individuals

country : United States
}
\references{
Verbeek, Marno (2004) A guide to modern econometrics, John Wiley and Sons, http://www.econ.kuleuven.ac.be/GME, chapter 7.

Journal of Business Economics and Statistics web site : http://www.amstat.org/publications/jbes/.
"benefits"
}
\keyword{datasets}
