% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\docType{data}
\name{blood_pressure}
\alias{blood_pressure}
\title{Blood Pressure}
\format{
A data frame
}
\usage{
blood_pressure
}
\description{
Systolic Blood pressure in mm/hg vs. daily saturated fat intake in mg for 25 subjects.
}
\keyword{datasets}
