% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\docType{data}
\name{car_use}
\alias{car_use}
\title{Car Use}
\format{
A data frame
}
\usage{
car_use
}
\description{
Car use, population and SO2 concentration in air in unidentified towns
}
\keyword{datasets}
