% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\docType{package}
\name{CodeClanData}
\alias{CodeClanData}
\alias{CodeClanData-package}
\title{Data for Data Analysis Course}
\description{
This package contains datasets for use with the Data Analysis course at Code Clan.
}
