% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\docType{data}
\name{colour_list}
\alias{colour_list}
\title{Colour List}
\format{
A nested list with 4 elements
}
\usage{
colour_list
}
\description{
A dataset with colour information (invented).
}
\keyword{datasets}
