% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\docType{data}
\name{competencies}
\alias{competencies}
\title{competencies}
\format{
A data frame
}
\usage{
competencies
}
\description{
Assessment of 4 widget companies' core competencies
}
\keyword{datasets}
