% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\docType{data}
\name{d20_outcomes}
\alias{d20_outcomes}
\title{d20 Outcomes}
\format{
A data frame
}
\usage{
d20_outcomes
}
\description{
The outcomes from rolling a 20 sided dice 201 times.
}
\keyword{datasets}
