% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\docType{data}
\name{d20x5_outcomes}
\alias{d20x5_outcomes}
\title{5d20 Outcomes}
\format{A data frame}
\usage{
d20x5_outcomes
}
\description{
The outcomes from rolling five 20 sided dice, and adding up the scores, 1001 times.
}
\keyword{datasets}
