% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\docType{data}
\name{drinks_content}
\alias{drinks_content}
\title{drinks_content}
\format{
A data frame
}
\usage{
drinks_content
}
\description{
Starbuck calorie dataset
https://www.kaggle.com/starbucks/starbucks-menu
}
\keyword{datasets}
