% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\docType{data}
\name{euro_ineq}
\alias{euro_ineq}
\title{euro_ineq}
\format{
A data frame
}
\usage{
euro_ineq
}
\description{
Volatility of GDP, Gini coefficient (measure of inequality) and percentage
of GDP attributable to services for 17 European countries.
}
\keyword{datasets}
