% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\docType{data}
\name{fitness_levels}
\alias{fitness_levels}
\title{Fitness Levels}
\format{
A data frame
}
\usage{
fitness_levels
}
\description{
Aerobic fitness levels of school children in different activity groups, over different ages.
}
\keyword{datasets}
