% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\docType{data}
\name{flatPrices}
\alias{flatPrices}
\title{flatPrices}
\format{
A data frame
}
\usage{
flatPrices
}
\description{
Estimated average flat prices and monthly rents in Edinburgh 2011-2019
}
\keyword{datasets}
