% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\docType{data}
\name{game_of_thrones}
\alias{game_of_thrones}
\title{Game of Thrones}
\format{A matrix}
\usage{
game_of_thrones
}
\description{
Data on about the book 'A Game of Thrones', including names
and genders for all the characters.
}
\keyword{datasets}
