% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\docType{data}
\name{game_sales}
\alias{game_sales}
\title{Video Game Sales Data}
\format{
A data frame
}
\usage{
game_sales
}
\description{
Global game sales data
}
\details{
name            : Title of the game

genre           : Type of game

year_of_release : Year the game was released

publisher       : Publisher of game

sales           : Total global sales, in millions

critic_score    : Critics review score (from 0 to 100) from Metacritic

user_score      : Users review score (from 0 to 10) from Metacritic

developer       : Studio that created the game

rating          : American ESRB rating for the game

platform        : Console the game runs on
}
\keyword{datasets}
