% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\docType{data}
\name{got_ratings}
\alias{got_ratings}
\title{Game of Thrones Ratings}
\format{
A data frame
}
\usage{
got_ratings
}
\description{
The IMDB ratings for Game of Thrones, season 1 to 8, by episode.
}
\keyword{datasets}
