% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\docType{data}
\name{heavily_right_skewed}
\alias{heavily_right_skewed}
\title{heavily right skewed data}
\format{
A data frame
}
\usage{
heavily_right_skewed
}
\description{
A dataset with simulated heavy right skewed data for teaching
}
\keyword{datasets}
