% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\docType{data}
\name{hospital_visits}
\alias{hospital_visits}
\title{hospital_visits}
\format{A data frame}
\usage{
hospital_visits
}
\description{
Reasons people were admitted to hospital between 1993 and 1998.
Purposely messy dataset
}
\keyword{datasets}
