% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\docType{data}
\name{inflation4}
\alias{inflation4}
\title{inflation4}
\format{A data frame}
\usage{
inflation4
}
\description{
Inflation (CPI) for UK, France, Germany and Spain 1960-2010
}
\keyword{datasets}
