% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\docType{data}
\name{invest_alluvial}
\alias{invest_alluvial}
\title{invest_alluvial}
\format{
A data frame
}
\usage{
invest_alluvial
}
\description{
Lifestyled pension investment allocations ages 40,50,60 - alluvial format
}
\keyword{datasets}
