% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\docType{data}
\name{invest_lodes}
\alias{invest_lodes}
\title{invest_lodes}
\format{
A data frame
}
\usage{
invest_lodes
}
\description{
Lifestyled pension investment allocations ages 40,50,60 - lodes format
}
\keyword{datasets}
