% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\docType{data}
\name{invest_lodes2}
\alias{invest_lodes2}
\title{invest_lodes2}
\format{A data frame}
\usage{
invest_lodes2
}
\description{
Lifestyled pension investment allocations ages 40,50,60 -
lodes format with outgoings
}
\keyword{datasets}
