% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\docType{data}
\name{iq_scores}
\alias{iq_scores}
\title{IQ Scores}
\format{
A data frame
}
\usage{
iq_scores
}
\description{
The IQ Scores for 5 people, measured across 3 different tests.
Invented data.
}
\keyword{datasets}
