% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\docType{data}
\name{kickstarter}
\alias{kickstarter}
\title{kickstarter}
\format{
A data frame
}
\usage{
kickstarter
}
\description{
Data for the crowd funding site Kickstarter 
https://www.kaggle.com/kemical/kickstarter-projects
}
\keyword{datasets}
