% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\docType{data}
\name{late_deliveries}
\alias{late_deliveries}
\title{late_deliveries}
\format{
A data frame
}
\usage{
late_deliveries
}
\description{
Number of late deliveries over 12 calendar months for fictional company
}
\keyword{datasets}
