% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\docType{data}
\name{left_skewed}
\alias{left_skewed}
\title{left skewed data}
\format{
A data frame
}
\usage{
left_skewed
}
\description{
A dataset with simulated left skewed data for teaching
}
\keyword{datasets}
