% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\docType{data}
\name{messy_orders}
\alias{messy_orders}
\title{customer orders}
\format{A data frame}
\usage{
messy_orders
}
\description{
Customer orders for three different people, with info and amount.
Invented data.
}
\keyword{datasets}
