% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\docType{data}
\name{monthly_sales}
\alias{monthly_sales}
\title{Monthly Sales}
\format{
A data frame
}
\usage{
monthly_sales
}
\description{
An invented dataset with monthly sales across 7 branches.
}
\details{
A summary of this dataset can be found in `total_sales`.
}
\keyword{datasets}
