% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\docType{data}
\name{new_coders}
\alias{new_coders}
\title{New Coders}
\format{
A data frame
}
\source{
\url{https://github.com/freeCodeCamp/2018-new-coder-survey}
}
\usage{
new_coders
}
\description{
A sample of data from freeCodeCamp's 2018 survey. Asks questions of new development
about their background and aspirations.
}
\keyword{datasets}
