% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\docType{data}
\name{nyc_dogs}
\alias{nyc_dogs}
\title{NYC Dogs}
\format{A data frame}
\usage{
nyc_dogs
}
\description{
A dataset with all the dogs registarted in New York City.
}
\details{
Adapted from: https://fusiontables.google.com/data?docid=1pKcxc8kzJbBVzLu_kgzoAMzqYhZyUhtScXjB0BQ#rows:id=1
}
\keyword{datasets}
