% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\docType{data}
\name{olympics_overall_medals}
\alias{olympics_overall_medals}
\title{Overall Olympics Medals}
\format{
A data frame
}
\usage{
olympics_overall_medals
}
\description{
The overall count of Olympic medals in both the summer and winter games.
From years 1896 to 2016.
Data from: https://www.kaggle.com/heesoo37/120-years-of-olympic-history-athletes-and-results/downloads/120-years-of-olympic-history-athletes-and-results.zip/2
}
\keyword{datasets}
