% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\docType{data}
\name{pension_liabilities}
\alias{pension_liabilities}
\title{Pension Liabilities}
\format{
A data frame
}
\usage{
pension_liabilities
}
\description{
Composition of liabilities (£m) of ABC pension scheme.
}
\keyword{datasets}
