% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\docType{data}
\name{playfair_denmark}
\alias{playfair_denmark}
\title{Playfair's Denmark Data}
\format{
A data frame
}
\usage{
playfair_denmark
}
\description{
Playfair's Denmark Data
}
\keyword{datasets}
