% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\docType{data}
\name{Q23data}
\alias{Q23data}
\title{QikBit_MonthlyFigs}
\format{A data frame}
\usage{
Q23data
}
\description{
Aug 2018 - Jul 2019 revenue and costs for QikBit, plus sales for single product line
}
\keyword{datasets}
