% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\docType{data}
\name{Q4data}
\alias{Q4data}
\title{QikBit_Competitors}
\format{A data frame}
\usage{
Q4data

Q4data
}
\description{
2015-2019 Revenue for QikBit's competitors

Comparative accelerometer data for 4 devices and 5 individuals over a single day
}
\keyword{datasets}
