% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\docType{data}
\name{qb_competitors}
\alias{qb_competitors}
\title{QikBit Competitors}
\format{A data frame}
\usage{
qb_competitors
}
\description{
2015-2019 Revenue for QikBit's competitors
}
\keyword{datasets}
