% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\docType{data}
\name{qb_device_data}
\alias{qb_device_data}
\title{QikBit Accelerometer Data}
\format{
A data frame
}
\usage{
qb_device_data
}
\description{
Comparative accelerometer data for 4 devices and 5 individuals over a single day
}
\keyword{datasets}
