% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\docType{data}
\name{qb_monthly_sales}
\alias{qb_monthly_sales}
\title{QikBit Monthly Figures}
\format{A data frame}
\usage{
qb_monthly_sales
}
\description{
Aug 2018 - Jul 2019 revenue and costs for QikBit, plus sales for single product line
}
\keyword{datasets}
