% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\docType{data}
\name{qb_revenue_breakdown}
\alias{qb_revenue_breakdown}
\title{QikBit Revenue Breakdown}
\format{A data frame}
\usage{
qb_revenue_breakdown
}
\description{
2019 Sales by product group for (fictional company) QikBit
}
\keyword{datasets}
