% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\docType{data}
\name{recovery_times}
\alias{recovery_times}
\title{Recovery times}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 4 rows and 4 columns.}
\usage{
recovery_times
}
\description{
Average Recovery Times under Treatments A and B, and Control.
}
\details{
@format A data frame
}
\keyword{datasets}
