% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\docType{data}
\name{refunds}
\alias{refunds}
\title{refunds}
\format{A data frame}
\usage{
refunds
}
\description{
Refunds for Brazils House of Deputies
https://www.kaggle.com/epattaro/brazils-house-of-deputies-reimbursements
}
\keyword{datasets}
