% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\docType{data}
\name{right_skewed}
\alias{right_skewed}
\title{right skewed data}
\format{
A data frame
}
\usage{
right_skewed
}
\description{
A dataset with simulated right skewed data for teaching
}
\keyword{datasets}
