% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\docType{data}
\name{savings_test}
\alias{savings_test}
\title{Savings - Test}
\format{A data frame}
\usage{
savings_test
}
\description{
Invented data about how much money people have saved.
}
\details{
Data for testing models.
}
\keyword{datasets}
