% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\docType{data}
\name{savings_train}
\alias{savings_train}
\title{Savings - Train}
\format{A data frame}
\usage{
savings_train
}
\description{
Invented data about how much money people have saved.
}
\details{
Data for training models.
}
\keyword{datasets}
