% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\docType{data}
\name{school_census}
\alias{school_census}
\title{school_census}
\format{A data frame}
\usage{
school_census
}
\description{
A New Zealand school census
https://new.censusatschool.org.nz/about/
}
\keyword{datasets}
