% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\docType{data}
\name{scot_exports_2017}
\alias{scot_exports_2017}
\title{scot_exports_2017}
\format{A data frame}
\usage{
scot_exports_2017
}
\description{
Scottish exports by sector (and sub-sector) in 2017
}
\keyword{datasets}
