% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\docType{data}
\name{starwars}
\alias{starwars}
\title{Starwars}
\format{
A list with 10 elements
}
\source{
\url{https://swapi.co/}
}
\usage{
starwars
}
\description{
An extract of the data available from the Star Wars API.
Information on 10 characters from Star Wars.
}
\keyword{datasets}
