% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\docType{data}
\name{state_income_data}
\alias{state_income_data}
\title{State Income Data}
\format{
A data frame
}
\usage{
state_income_data
}
\description{
A dataset with information about incomes in states.
}
\keyword{datasets}
