% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\docType{data}
\name{students_big}
\alias{students_big}
\title{Students Big}
\format{A data frame}
\source{
\url{https://new.censusatschool.org.nz/tools/random-sampler/}
}
\usage{
students_big
}
\description{
A larger dataset with information from the student census. This time only for UK students.
}
\keyword{datasets}
