% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\docType{data}
\name{table_of_numbers}
\alias{table_of_numbers}
\title{Table of Numbers}
\format{A data frame}
\usage{
table_of_numbers
}
\description{
A randomly generated table of 100 numbers. Numbers come from a Poisson distribution
with mean 3.
}
\keyword{datasets}
