% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\docType{data}
\name{temp}
\alias{temp}
\title{Temperature}
\format{
A matrix
}
\usage{
temp
}
\description{
A matrix which contains the maximum temperature for each month in Scotland
over 106 years (1910 from 2015). Each row corresponds to a year and each
column to a month (January to December).
}
\keyword{datasets}
