% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\docType{data}
\name{total_sales}
\alias{total_sales}
\title{Total sales}
\format{
A data frame
}
\usage{
total_sales
}
\description{
An invented dataset with total sales across 7 branches.
}
\details{
A monthly breakdown of this dataset can be found in `monthly_sales`.
}
\keyword{datasets}
