% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\docType{data}
\name{vote19_eu}
\alias{vote19_eu}
\title{VoteEU}
\format{A data frame}
\usage{
vote19_eu
}
\description{
Two way table - vote in 2016 EU referendum and 2019 voting intention
Converted to ggplot2 layout
}
\keyword{datasets}
