% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\docType{data}
\name{women_in_gov}
\alias{women_in_gov}
\title{women_in_gov}
\format{
A data frame
}
\usage{
women_in_gov
}
\description{
Women in Parliament
https://data.worldbank.org/indicator/SG.GEN.PARL.ZS
}
\keyword{datasets}
