% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\docType{data}
\name{students}
\alias{students}
\title{Students}
\format{A data frame with 14 observations of 10 variables}
\source{
\url{https://new.censusatschool.org.nz/tools/random-sampler/}
}
\usage{
students
}
\description{
A small dataset with some survey information from randomly chosen students.
Comes from the international census at school random sampler.
}
\keyword{datasets}
