% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\docType{data}
\name{temp_df}
\alias{temp_df}
\title{Temperature Data Frame}
\format{A data frame}
\usage{
temp_df
}
\description{
The same data available in temp, but in a long data frame.
}
\keyword{datasets}
